requirements.txt
google-generativeai==0.8.4
streamlit==1.30.0
pillow==11.1.0
streamlit-option-menu==0.3.12